add entitlement for installing system extension
