+++ electron-builder/entitlements.plist
@@ -14,5 +14,7 @@
     <true/>
     <key>com.apple.security.cs.allow-jit</key>
     <true/>
+    <key>com.apple.developer.system-extension.install</key>
+    <true/>
   </dict>
 </plist>
